# apps/api/trader-assistant/trading-dashboard/openAPI.yaml
openapi: 3.0.3
info:
  title: Trading Dashboard API
  version: 1.0.0
  description: |
    Contract for the Benji trader assistant API. Endpoints expose quote intelligence,
    portfolio holdings, and personal finance insights consumed by the dashboard UI.
security:
  - ApiKeyAuth: []
paths:
  /api/quotes:
    get:
      tags:
        - Quotes
      operationId: getQuotesIndex
      summary: Describe the quote service entry point.
      responses:
        '200':
          description: Quote service metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotesIndex'
  /api/quotes/{symbol}:
    get:
      tags:
        - Quotes
      operationId: getQuote
      summary: Retrieve the latest quote for a ticker symbol.
      parameters:
        - name: symbol
          in: path
          required: true
          description: Equity or ETF ticker symbol (case insensitive).
          schema:
            type: string
            minLength: 1
            maxLength: 12
      responses:
        '200':
          description: Quote payload for the requested symbol.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponse'
        '400':
          description: Invalid ticker was provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No quote could be located for the symbol.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/portfolio/positions:
    get:
      tags:
        - Portfolio
      operationId: listPortfolioPositions
      summary: List normalized portfolio positions and cost basis.
      responses:
        '200':
          description: Portfolio positions currently stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioPositionsResponse'
        '204':
          description: No portfolio positions available.
    post:
      tags:
        - Portfolio
      operationId: createPortfolioPosition
      summary: Create a new portfolio position.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePortfolioPositionRequest'
      responses:
        '201':
          description: Position created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioPosition'
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Position for ticker already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/portfolio/positions/{ticker}:
    get:
      tags:
        - Portfolio
      operationId: getPortfolioPosition
      summary: Retrieve a single portfolio position by ticker.
      parameters:
        - name: ticker
          in: path
          required: true
          description: Stock ticker symbol.
          schema:
            type: string
            minLength: 1
            maxLength: 12
      responses:
        '200':
          description: Position found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioPosition'
        '404':
          description: Position not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Portfolio
      operationId: updatePortfolioPosition
      summary: Update an existing portfolio position.
      parameters:
        - name: ticker
          in: path
          required: true
          description: Stock ticker symbol.
          schema:
            type: string
            minLength: 1
            maxLength: 12
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePortfolioPositionRequest'
      responses:
        '200':
          description: Position updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioPosition'
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Position not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Portfolio
      operationId: deletePortfolioPosition
      summary: Delete a portfolio position.
      parameters:
        - name: ticker
          in: path
          required: true
          description: Stock ticker symbol.
          schema:
            type: string
            minLength: 1
            maxLength: 12
      responses:
        '204':
          description: Position deleted successfully.
        '404':
          description: Position not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/portfolio/summary:
    get:
      tags:
        - Portfolio
      operationId: getPortfolioSummary
      summary: Provide aggregate portfolio metrics derived from stored positions.
      responses:
        '200':
          description: Portfolio summary view.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioSummary'
        '204':
          description: No portfolio data is available to summarize.
  /api/finance/summary:
    get:
      tags:
        - Finance
      operationId: getFinanceSummary
      summary: Provide high-level spend analytics for the current month.
      responses:
        '200':
          description: Finance snapshot derived from recent transactions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceSummary'
        '204':
          description: No finance data is available.
  /api/finance/transactions:
    get:
      tags:
        - Finance
      operationId: listFinanceTransactions
      summary: List categorized personal finance transactions.
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of most recent transactions to return.
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: category
          in: query
          required: false
          description: Filter transactions by normalized category slug (e.g., housing, dining).
          schema:
            type: string
      responses:
        '200':
          description: Finance transactions matching the criteria.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceTransactionsResponse'
        '204':
          description: No transactions matched the query.
    post:
      tags:
        - Finance
      operationId: createFinanceTransaction
      summary: Create a new finance transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFinanceTransactionRequest'
      responses:
        '201':
          description: Transaction created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceTransaction'
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/finance/transactions/{id}:
    get:
      tags:
        - Finance
      operationId: getFinanceTransaction
      summary: Retrieve a single finance transaction by ID.
      parameters:
        - name: id
          in: path
          required: true
          description: Unique transaction identifier.
          schema:
            type: string
      responses:
        '200':
          description: Transaction found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceTransaction'
        '404':
          description: Transaction not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Finance
      operationId: updateFinanceTransaction
      summary: Update an existing finance transaction.
      parameters:
        - name: id
          in: path
          required: true
          description: Unique transaction identifier.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFinanceTransactionRequest'
      responses:
        '200':
          description: Transaction updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceTransaction'
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transaction not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Finance
      operationId: deleteFinanceTransaction
      summary: Delete a finance transaction.
      parameters:
        - name: id
          in: path
          required: true
          description: Unique transaction identifier.
          schema:
            type: string
      responses:
        '204':
          description: Transaction deleted successfully.
        '404':
          description: Transaction not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    QuotesIndex:
      type: object
      required:
        - message
        - endpoints
      properties:
        message:
          type: string
          example: Quote service ready.
        endpoints:
          type: array
          minItems: 1
          description: Helpful endpoints exposed by the quote service.
          items:
            type: string
            example: /api/quotes/{symbol}
    QuoteResponse:
      type: object
      required:
        - symbol
        - price
        - as_of
      properties:
        symbol:
          type: string
          description: Requested ticker symbol.
          example: AAPL
        price:
          type: number
          format: double
          description: Last traded price reported by the primary provider.
          example: 193.63
        currency:
          type: string
          description: ISO 4217 currency code for the quote.
          example: USD
        as_of:
          type: string
          format: date-time
          description: Timestamp the quote was last refreshed.
          example: '2024-07-19T15:30:00Z'
        source:
          type: string
          description: Market data provider that supplied the quote.
          example: AlphaVantage
    PortfolioPositionsResponse:
      type: object
      required:
        - positions
      properties:
        as_of:
          type: string
          format: date-time
          description: Snapshot timestamp when the positions were generated.
        positions:
          type: array
          description: Portfolio positions persisted by the ingestion pipeline.
          items:
            $ref: '#/components/schemas/PortfolioPosition'
    PortfolioPosition:
      type: object
      required:
        - ticker
        - quantity
        - cost_basis
      properties:
        ticker:
          type: string
          description: Normalized ticker symbol.
          example: VOO
        quantity:
          type: number
          format: double
          description: Total shares held for the ticker.
          example: 12.5
        cost_basis:
          type: number
          format: double
          description: Aggregate cost basis for the position.
          example: 4500.12
    PortfolioSummary:
      type: object
      required:
        - positions_count
        - total_quantity
        - total_cost_basis
      properties:
        positions_count:
          type: integer
          format: int32
          description: Number of unique tickers in the portfolio.
          example: 8
        total_quantity:
          type: number
          format: double
          description: Sum of share quantities across all positions.
          example: 154.72
        total_cost_basis:
          type: number
          format: double
          description: Total invested cost basis across the portfolio.
          example: 25340.55
        as_of:
          type: string
          format: date-time
          description: Timestamp when the summary metrics were calculated.
    FinanceSummary:
      type: object
      required:
        - month_to_date_spend
        - average_daily_spend
        - projected_month_end_spend
        - as_of
      properties:
        month_to_date_spend:
          type: number
          format: double
          description: Total spend recorded for the current calendar month.
          example: 2845.22
        average_daily_spend:
          type: number
          format: double
          description: Average daily spend computed for the active month.
          example: 142.26
        projected_month_end_spend:
          type: number
          format: double
          description: Forecasted spend by the end of the month.
          example: 4400.0
        as_of:
          type: string
          format: date-time
          description: Timestamp when analytics were calculated.
    FinanceTransactionsResponse:
      type: object
      required:
        - transactions
      properties:
        as_of:
          type: string
          format: date-time
          description: Snapshot timestamp for the transaction list.
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/FinanceTransaction'
          description: Transactions ordered from newest to oldest.
    FinanceTransaction:
      type: object
      required:
        - id
        - posted_at
        - description
        - amount
        - category
      properties:
        id:
          type: string
          description: Unique identifier for the transaction.
          example: txn_20240719_001
        posted_at:
          type: string
          format: date-time
          description: Timestamp the transaction posted.
          example: '2024-07-18T12:00:00Z'
        description:
          type: string
          description: Human readable description.
          example: Whole Foods Market
        amount:
          type: number
          format: double
          description: Signed transaction amount; debits are negative.
          example: -72.34
        category:
          type: string
          description: Normalized category slug.
          example: groceries
        notes:
          type: string
          description: Optional annotations captured by the user.
    CreateFinanceTransactionRequest:
      type: object
      required:
        - posted_at
        - description
        - amount
        - category
      properties:
        posted_at:
          type: string
          format: date-time
          description: Timestamp the transaction posted.
          example: '2024-07-18T12:00:00Z'
        description:
          type: string
          minLength: 1
          maxLength: 255
          description: Human readable description.
          example: Whole Foods Market
        amount:
          type: number
          format: double
          description: Signed transaction amount; debits are negative.
          example: -72.34
        category:
          type: string
          minLength: 1
          maxLength: 64
          description: Normalized category slug.
          example: groceries
        notes:
          type: string
          maxLength: 1000
          description: Optional annotations captured by the user.
    UpdateFinanceTransactionRequest:
      type: object
      properties:
        posted_at:
          type: string
          format: date-time
          description: Timestamp the transaction posted.
          example: '2024-07-18T12:00:00Z'
        description:
          type: string
          minLength: 1
          maxLength: 255
          description: Human readable description.
          example: Whole Foods Market
        amount:
          type: number
          format: double
          description: Signed transaction amount; debits are negative.
          example: -72.34
        category:
          type: string
          minLength: 1
          maxLength: 64
          description: Normalized category slug.
          example: groceries
        notes:
          type: string
          maxLength: 1000
          description: Optional annotations captured by the user.
    CreatePortfolioPositionRequest:
      type: object
      required:
        - ticker
        - quantity
        - cost_basis
      properties:
        ticker:
          type: string
          minLength: 1
          maxLength: 12
          pattern: '^[A-Za-z0-9.]+$'
          description: Stock ticker symbol.
          example: VOO
        quantity:
          type: number
          format: double
          minimum: 0
          description: Total shares held for the ticker.
          example: 12.5
        cost_basis:
          type: number
          format: double
          minimum: 0
          description: Aggregate cost basis for the position.
          example: 4500.12
    UpdatePortfolioPositionRequest:
      type: object
      properties:
        quantity:
          type: number
          format: double
          minimum: 0
          description: Total shares held for the ticker.
          example: 12.5
        cost_basis:
          type: number
          format: double
          minimum: 0
          description: Aggregate cost basis for the position.
          example: 4500.12
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine readable error code.
          example: SYMBOL_INVALID
        message:
          type: string
          description: Explanation of the error condition.
          example: Provided ticker symbol is invalid.
        details:
          type: array
          description: Additional detail strings.
          items:
            type: string
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: API key for authenticating requests to the Trading Dashboard API.